pyproject.toml
[tool.poetry]
name = "anti-spam-bot"
version = "0.1.0"
description = ""
authors = ["singapore9 <sharik.and@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = {version = "20.0a2", allow-prereleases = true}
fastapi = "^0.85.0"
requests = "^2.28.1"
firebase-admin = "^5.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
